Refactor the Worker logic to Offline and Online tasks
## Summary:

This refactor structure is pretty much personal taste here.
I factor logic loop to smaller chunk, along the way I also improve the TaskFetcher so it does not give up early `MAX404_RETRY`.

This PR improved #1628. So it should be merge together.
I made the changes after refactor so I guess it's hard separate the improvement and the refactor together.

## Before

```
% tree .
.

├── prover_runtime.rs
├── prover.rs
├── register.rs
├── system.rs
├── task_cache.rs
├── task.rs
└── ui
    ├── dashboard.rs
    ├── login.rs
    ├── mod.rs
    └── splash.rs
```

## After:

```
% tree .
.
├── prover_runtime.rs
├── prover.rs
├── register.rs
├── system.rs
├── task_cache.rs
├── task.rs
├── ui
│   ├── dashboard.rs
│   ├── login.rs
│   ├── mod.rs
│   └── splash.rs
└── workers
    ├── mod.rs
    ├── offline.rs
    └── online.rs
```
