+++ clients/cli/src/prover_runtime.rs
@@ -1,144 +1,18 @@
 //! Prover Runtime
 //!
-//! Handles background execution of proof tasks in both authenticated and anonymous modes.
-//! Spawns async workers, dispatches tasks, and reports progress back to the UI.
-//!
-//! Includes:
-//! - Task fetching from the orchestrator (authenticated mode)
-//! - Worker management and task dispatching
-//! - Prover event reporting
+//! Main orchestrator for authenticated and anonymous proving modes.
+//! Coordinates online workers (network I/O) and offline workers (computation).
 
 use crate::environment::Environment;
-use crate::error_classifier::{ErrorClassifier, LogLevel};
-use crate::logging::should_log_with_env;
-use crate::orchestrator::error::OrchestratorError;
-use crate::orchestrator::{Orchestrator, OrchestratorClient};
-use crate::prover::authenticated_proving;
+use crate::events::Event;
+use crate::orchestrator::OrchestratorClient;
 use crate::task::Task;
 use crate::task_cache::TaskCache;
-use chrono::Local;
-use ed25519_dalek::{SigningKey, VerifyingKey};
+use crate::workers::{offline, online};
+use ed25519_dalek::SigningKey;
 use nexus_sdk::stwo::seq::Proof;
-use sha3::{Digest, Keccak256};
-use std::fmt::Display;
-use std::string::ToString;
-use std::time::Duration;
 use tokio::sync::{broadcast, mpsc};
 use tokio::task::JoinHandle;
-
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-pub enum Worker {
-    /// Worker that fetches tasks from the orchestrator and processes them.
-    TaskFetcher,
-    /// Worker that performs proving tasks.
-    Prover(usize),
-    /// Worker that submits proofs to the orchestrator.
-    ProofSubmitter,
-}
-
-#[derive(Debug, Copy, Clone, Eq, PartialEq, strum::Display)]
-pub enum EventType {
-    Success,
-    Error,
-    Refresh,
-    Shutdown,
-}
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct Event {
-    pub worker: Worker,
-    pub msg: String,
-    pub timestamp: String,
-    pub event_type: EventType,
-    pub log_level: LogLevel,
-}
-
-impl Event {
-    pub fn new(kind: Worker, msg: String, event_type: EventType) -> Self {
-        Self {
-            worker: kind,
-            msg,
-            timestamp: Local::now().format("%Y-%m-%d %H:%M:%S").to_string(),
-            event_type,
-            log_level: LogLevel::Info,
-        }
-    }
-
-    pub fn new_with_level(
-        kind: Worker,
-        msg: String,
-        event_type: EventType,
-        log_level: LogLevel,
-    ) -> Self {
-        Self {
-            worker: kind,
-            msg,
-            timestamp: Local::now().format("%Y-%m-%d %H:%M:%S").to_string(),
-            event_type,
-            log_level,
-        }
-    }
-
-    pub fn task_fetcher(msg: String, event_type: EventType) -> Self {
-        Self::new(Worker::TaskFetcher, msg, event_type)
-    }
-
-    pub fn task_fetcher_with_level(
-        msg: String,
-        event_type: EventType,
-        log_level: LogLevel,
-    ) -> Self {
-        Self::new_with_level(Worker::TaskFetcher, msg, event_type, log_level)
-    }
-
-    pub fn prover(worker_id: usize, msg: String, event_type: EventType) -> Self {
-        Self::new(Worker::Prover(worker_id), msg, event_type)
-    }
-
-    pub fn prover_with_level(
-        worker_id: usize,
-        msg: String,
-        event_type: EventType,
-        log_level: LogLevel,
-    ) -> Self {
-        Self::new_with_level(Worker::Prover(worker_id), msg, event_type, log_level)
-    }
-
-    pub fn proof_submitter(msg: String, event_type: EventType) -> Self {
-        Self::new(Worker::ProofSubmitter, msg, event_type)
-    }
-
-    pub fn proof_submitter_with_level(
-        msg: String,
-        event_type: EventType,
-        log_level: LogLevel,
-    ) -> Self {
-        Self::new_with_level(Worker::ProofSubmitter, msg, event_type, log_level)
-    }
-
-    pub fn should_display(&self) -> bool {
-        // Always show success events and info level events
-        if self.event_type == EventType::Success || self.log_level >= LogLevel::Info {
-            return true;
-        }
-        should_log_with_env(self.log_level)
-    }
-}
-
-impl Display for Event {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let worker_type: String = match self.worker {
-            Worker::TaskFetcher => "Task Fetcher".to_string(),
-            Worker::Prover(worker_id) => format!("Prover {}", worker_id),
-            Worker::ProofSubmitter => "Proof Submitter".to_string(),
-        };
-        write!(
-            f,
-            "{} [{}] {}: {}",
-            self.event_type, self.timestamp, worker_type, self.msg
-        )
-    }
-}
 
 // Queue sizes. Chosen to be larger than the tasks API page size (currently, 50)
 const TASK_QUEUE_SIZE: usize = 100;
@@ -147,10 +21,6 @@ const RESULT_QUEUE_SIZE: usize = 100;
 
 /// Maximum number of completed tasks to keep in memory. Chosen to be larger than the task queue size.
 const MAX_COMPLETED_TASKS: usize = 500;
-
-// Task fetching thresholds
-const LOW_WATER_MARK: usize = 5; // Fetch new tasks when queue drops below this
-const BATCH_SIZE: usize = 20; // Fetch this many tasks at once
 
 /// Starts authenticated workers that fetch tasks from the orchestrator and process them.
 pub async fn start_authenticated_workers(
@@ -177,7 +47,7 @@ pub async fn start_authenticated_workers(
         let event_sender = event_sender.clone();
         let shutdown = shutdown.resubscribe(); // Clone the receiver for task fetching
         tokio::spawn(async move {
-            fetch_prover_tasks(
+            online::fetch_prover_tasks(
                 node_id,
                 verifying_key,
                 Box::new(orchestrator),
@@ -194,7 +64,7 @@ pub async fn start_authenticated_workers(
     // Workers
     let (result_sender, result_receiver) = mpsc::channel::<(Task, Proof)>(RESULT_QUEUE_SIZE);
 
-    let (worker_senders, worker_handles) = start_workers(
+    let (worker_senders, worker_handles) = offline::start_workers(
         num_workers,
         result_sender,
         event_sender.clone(),
@@ -205,14 +75,15 @@ pub async fn start_authenticated_workers(
     join_handles.extend(worker_handles);
 
     // Dispatch tasks to workers
-    let dispatcher_handle = start_dispatcher(task_receiver, worker_senders, shutdown.resubscribe());
+    let dispatcher_handle =
+        offline::start_dispatcher(task_receiver, worker_senders, shutdown.resubscribe());
     join_handles.push(dispatcher_handle);
 
     // A bounded list of recently completed task IDs
     let successful_tasks = TaskCache::new(MAX_COMPLETED_TASKS);
 
     // Send proofs to the orchestrator
-    let submit_proofs_handle = submit_proofs(
+    let submit_proofs_handle = online::submit_proofs(
         signing_key,
         Box::new(orchestrator),
         num_workers,
@@ -234,522 +105,13 @@ pub async fn start_anonymous_workers(
     environment: Environment,
     client_id: String,
 ) -> (mpsc::Receiver<Event>, Vec<JoinHandle<()>>) {
-    let (event_sender, event_receiver) = mpsc::channel::<Event>(100);
-    let mut join_handles = Vec::new();
-    for worker_id in 0..num_workers {
-        let prover_event_sender = event_sender.clone();
-        let mut shutdown_rx = shutdown.resubscribe(); // clone receiver for each worker
-        let client_id = client_id.clone();
-        let error_classifier = ErrorClassifier::new();
-
-        let handle = tokio::spawn(async move {
-            loop {
-                tokio::select! {
-                    _ = shutdown_rx.recv() => {
-                        let message = format!("Worker {} received shutdown signal", worker_id);
-                        let _ = prover_event_sender
-                            .send(Event::prover(worker_id, message, EventType::Shutdown))
-                            .await;
-                        break; // Exit the loop on shutdown signal
-                    }
-
-                    _ = tokio::time::sleep(Duration::from_millis(300)) => {
-                        // Perform work
-                        match crate::prover::prove_anonymously(&environment, client_id.clone()).await {
-                            Ok(_proof) => {
-                                let message = "Anonymous proof completed successfully".to_string();
-                                let _ = prover_event_sender
-                                    .send(Event::prover(worker_id, message, EventType::Success)).await;
-                            }
-                            Err(e) => {
-                                let log_level = error_classifier.classify_worker_error(&e);
-                                let message = format!("Anonymous Worker: Error - {}", e);
-                                let event = Event::prover_with_level(worker_id, message, EventType::Error, log_level);
-                                if event.should_display() {
-                                    let _ = prover_event_sender.send(event).await;
-                                }
-
-                                // For analytics errors, this is non-critical, continue the loop
-                                // For other errors, also continue (anonymous mode keeps retrying)
-                            }
-                        }
-                    }
-                }
-            }
-        });
-        join_handles.push(handle);
-    }
-
-    (event_receiver, join_handles)
-}
-
-/// State for managing task fetching behavior
-struct TaskFetchState {
-    last_fetch_time: std::time::Instant,
-    backoff_duration: Duration,
-    last_queue_log_time: std::time::Instant,
-    queue_log_interval: Duration,
-    error_classifier: ErrorClassifier,
-}
-
-impl TaskFetchState {
-    fn new() -> Self {
-        Self {
-            last_fetch_time: std::time::Instant::now() - Duration::from_millis(31000), // Allow immediate first fetch
-            backoff_duration: Duration::from_millis(30000), // Start with 30 second backoff
-            last_queue_log_time: std::time::Instant::now(),
-            queue_log_interval: Duration::from_millis(60000), // Log queue status every 1 minute max
-            error_classifier: ErrorClassifier::new(),
-        }
-    }
-
-    fn should_log_queue_status(&mut self, tasks_in_queue: usize) -> bool {
-        tasks_in_queue < LOW_WATER_MARK
-            && self.last_queue_log_time.elapsed() >= self.queue_log_interval
-    }
-
-    fn should_fetch(&self, tasks_in_queue: usize) -> bool {
-        tasks_in_queue < LOW_WATER_MARK && self.last_fetch_time.elapsed() >= self.backoff_duration
-    }
-
-    fn record_fetch_attempt(&mut self) {
-        self.last_fetch_time = std::time::Instant::now();
-    }
-
-    fn record_queue_log(&mut self) {
-        self.last_queue_log_time = std::time::Instant::now();
-    }
-
-    fn reset_backoff(&mut self) {
-        self.backoff_duration = Duration::from_millis(30000);
-    }
-
-    fn increase_backoff_for_rate_limit(&mut self) {
-        self.backoff_duration =
-            std::cmp::min(self.backoff_duration * 2, Duration::from_millis(60000));
-    }
-
-    fn increase_backoff_for_error(&mut self) {
-        self.backoff_duration =
-            std::cmp::min(self.backoff_duration * 2, Duration::from_millis(30000));
-    }
-}
-
-/// Fetches tasks from the orchestrator and place them in the task queue.
-/// Uses demand-driven fetching: only fetches when queue drops below LOW_WATER_MARK.
-pub async fn fetch_prover_tasks(
-    node_id: u64,
-    verifying_key: VerifyingKey,
-    orchestrator_client: Box<dyn Orchestrator>,
-    sender: mpsc::Sender<Task>,
-    event_sender: mpsc::Sender<Event>,
-    mut shutdown: broadcast::Receiver<()>,
-    recent_tasks: TaskCache,
-) {
-    let mut state = TaskFetchState::new();
-
-    loop {
-        tokio::select! {
-            _ = shutdown.recv() => break,
-            _ = tokio::time::sleep(Duration::from_millis(500)) => {
-                let tasks_in_queue = TASK_QUEUE_SIZE - sender.capacity();
-
-                // Log queue status occasionally
-                if state.should_log_queue_status(tasks_in_queue) {
-                    state.record_queue_log();
-                    log_queue_status(&event_sender, tasks_in_queue, &state).await;
-                }
-
-                // Attempt fetch if conditions are met
-                if state.should_fetch(tasks_in_queue) {
-                    state.record_fetch_attempt();
-
-                    match fetch_task_batch(&*orchestrator_client, &node_id, verifying_key, BATCH_SIZE).await {
-                        Ok(tasks) => {
-                            if let Err(should_return) = handle_fetch_success(
-                                tasks,
-                                &sender,
-                                &event_sender,
-                                &recent_tasks,
-                                &mut state
-                            ).await {
-                                if should_return {
-                                    return;
-                                }
-                            }
-                        }
-                        Err(e) => {
-                            handle_fetch_error(e, &event_sender, &mut state).await;
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Log the current queue status
-async fn log_queue_status(
-    event_sender: &mpsc::Sender<Event>,
-    tasks_in_queue: usize,
-    state: &TaskFetchState,
-) {
-    let time_since_last = state.last_fetch_time.elapsed();
-    let backoff_secs = state.backoff_duration.as_secs();
-
-    let message = if time_since_last >= state.backoff_duration {
-        format!("Queue low: {} tasks, fetching now...", tasks_in_queue)
-    } else {
-        let time_since_secs = time_since_last.as_secs();
-        format!(
-            "Queue low: {} tasks, last fetch {}s ago (retry every {}s)",
-            tasks_in_queue, time_since_secs, backoff_secs
-        )
-    };
-
-    let _ = event_sender
-        .send(Event::task_fetcher_with_level(
-            message,
-            EventType::Refresh,
-            LogLevel::Debug,
-        ))
-        .await;
-}
-
-/// Handle successful task fetch
-async fn handle_fetch_success(
-    tasks: Vec<Task>,
-    sender: &mpsc::Sender<Task>,
-    event_sender: &mpsc::Sender<Event>,
-    recent_tasks: &TaskCache,
-    state: &mut TaskFetchState,
-) -> Result<(), bool> {
-    // bool indicates if caller should return
-    if tasks.is_empty() {
-        let _ = event_sender
-            .send(Event::task_fetcher_with_level(
-                "No tasks available from server".to_string(),
-                EventType::Refresh,
-                LogLevel::Info,
-            ))
-            .await;
-        return Ok(());
-    }
-
-    let mut added_count = 0;
-    for task in tasks {
-        if recent_tasks.contains(&task.task_id).await {
-            continue;
-        }
-        recent_tasks.insert(task.task_id.clone()).await;
-
-        if sender.send(task).await.is_err() {
-            let _ = event_sender
-                .send(Event::task_fetcher(
-                    "Task queue is closed".to_string(),
-                    EventType::Shutdown,
-                ))
-                .await;
-            return Err(true); // Signal caller to return
-        }
-        added_count += 1;
-    }
-
-    if added_count > 0 {
-        // Only log significant additions
-        if added_count >= 5 {
-            let msg = format!(
-                "Added {} tasks to queue (queue level: {})",
-                added_count,
-                TASK_QUEUE_SIZE - sender.capacity()
-            );
-            let _ = event_sender
-                .send(Event::task_fetcher_with_level(
-                    msg,
-                    EventType::Refresh,
-                    LogLevel::Info,
-                ))
-                .await;
-        }
-        state.reset_backoff();
-    }
-
-    Ok(())
-}
-
-/// Handle fetch errors with appropriate backoff
-async fn handle_fetch_error(
-    error: OrchestratorError,
-    event_sender: &mpsc::Sender<Event>,
-    state: &mut TaskFetchState,
-) {
-    if matches!(error, OrchestratorError::Http { status: 429, .. }) {
-        state.increase_backoff_for_rate_limit();
-        let _ = event_sender
-            .send(Event::task_fetcher_with_level(
-                format!(
-                    "Rate limited (429), backing off for {} seconds",
-                    state.backoff_duration.as_secs()
-                ),
-                EventType::Error,
-                LogLevel::Warn,
-            ))
-            .await;
-    } else {
-        state.increase_backoff_for_error();
-        let log_level = state.error_classifier.classify_fetch_error(&error);
-        let event = Event::task_fetcher_with_level(
-            format!(
-                "Failed to fetch tasks: {}, retrying in {} seconds",
-                error,
-                state.backoff_duration.as_secs()
-            ),
-            EventType::Error,
-            log_level,
-        );
-        if event.should_display() {
-            let _ = event_sender.send(event).await;
-        }
-    }
-}
-
-/// Fetch a batch of tasks from the orchestrator
-async fn fetch_task_batch(
-    orchestrator_client: &dyn Orchestrator,
-    node_id: &u64,
-    verifying_key: VerifyingKey,
-    batch_size: usize,
-) -> Result<Vec<Task>, OrchestratorError> {
-    // First try to get existing assigned tasks
-    match orchestrator_client.get_tasks(&node_id.to_string()).await {
-        Ok(tasks) => {
-            if !tasks.is_empty() {
-                return Ok(tasks);
-            }
-        }
-        Err(e) => {
-            // If getting existing tasks fails, try to get new ones
-            if !matches!(e, OrchestratorError::Http { status: 404, .. }) {
-                return Err(e);
-            }
-        }
-    }
-
-    // If no existing tasks, try to get new ones
-    let mut new_tasks = Vec::new();
-    for _ in 0..batch_size {
-        match orchestrator_client
-            .get_proof_task(&node_id.to_string(), verifying_key)
-            .await
-        {
-            Ok(task) => new_tasks.push(task),
-            Err(OrchestratorError::Http { status: 429, .. }) => {
-                // Rate limited, return what we have
-                break;
-            }
-            Err(OrchestratorError::Http { status: 404, .. }) => {
-                // No more tasks available
-                break;
-            }
-            Err(e) => return Err(e),
-        }
-    }
-
-    Ok(new_tasks)
-}
-
-/// Submits proofs to the orchestrator
-pub async fn submit_proofs(
-    signing_key: SigningKey,
-    orchestrator: Box<dyn Orchestrator>,
-    num_workers: usize,
-    mut results: mpsc::Receiver<(Task, Proof)>,
-    event_sender: mpsc::Sender<Event>,
-    mut shutdown: broadcast::Receiver<()>,
-    successful_tasks: TaskCache,
-) -> JoinHandle<()> {
-    tokio::spawn(async move {
-        loop {
-            tokio::select! {
-                maybe_item = results.recv() => {
-                    match maybe_item {
-                        Some((task, proof)) => {
-                            if successful_tasks.contains(&task.task_id).await {
-                                let msg = format!(
-                                        "Ignoring proof for previously submitted task {}",
-                                        task.task_id
-                                    );
-                                    let _ = event_sender
-                                        .send(Event::proof_submitter(msg, EventType::Error))
-                                        .await;
-                                // Skip already completed tasks
-                                continue;
-                            }
-                            let proof_bytes = postcard::to_allocvec(&proof)
-                                .expect("Failed to serialize proof");
-                            let proof_hash = format!("{:x}", Keccak256::digest(&proof_bytes));
-                            // let now = Local::now();
-                            // let timestamp = now.format("%Y-%m-%d %H:%M:%S").to_string();
-                            match orchestrator
-                                .submit_proof(&task.task_id, &proof_hash, proof_bytes, signing_key.clone(), num_workers)
-                                .await
-                            {
-                                Ok(_) => {
-                                    // Mark task as completed
-                                    successful_tasks.insert(task.task_id.clone()).await;
-                                    let msg = format!(
-                                        "Successfully submitted proof for task {}",
-                                        task.task_id
-                                    );
-                                    let _ = event_sender
-                                        .send(Event::proof_submitter_with_level(msg, EventType::Success, LogLevel::Info))
-                                        .await;
-                            }
-                                Err(OrchestratorError::Http {status, message: _message}) => {
-                                    let msg = format!(
-                                        "Failed to submit proof for task {}. Status: {}",
-                                        task.task_id,
-                                        status,
-                                    );
-                                    let _ = event_sender
-                                        .send(Event::proof_submitter(msg, EventType::Error))
-                                        .await;
-                                }
-                                Err(e) => {
-                                    let msg = format!(
-                                        "Failed to submit proof for task {}: {}",
-                                        task.task_id,
-                                        e
-                                    );
-                                    let _ = event_sender
-                                        .send(Event::proof_submitter(msg, EventType::Error))
-                                        .await;
-                                }
-                            }
-                        }
-                        None => {
-                            break;
-                        }
-                    }
-                }
-
-                _ = shutdown.recv() => {
-                    break;
-                }
-            }
-        }
-    })
-}
-
-/// Spawns a dispatcher that forwards tasks to available workers in round-robin fashion.
-pub fn start_dispatcher(
-    mut task_receiver: mpsc::Receiver<Task>,
-    worker_senders: Vec<mpsc::Sender<Task>>,
-    mut shutdown: broadcast::Receiver<()>,
-) -> JoinHandle<()> {
-    tokio::spawn(async move {
-        let mut next_worker = 0;
-        loop {
-            tokio::select! {
-                Some(task) = task_receiver.recv() => {
-                    let target = next_worker % worker_senders.len();
-                    if let Err(_e) = worker_senders[target].send(task).await {
-                        // Channel is closed, stop dispatching tasks
-                        return;
-                    }
-                    next_worker += 1;
-                }
-
-                _ = shutdown.recv() => {
-                    break;
-                }
-            }
-        }
-    })
-}
-
-/// Spawns a set of worker tasks that receive tasks and send prover events.
-///
-/// # Arguments
-/// * `num_workers` - The number of worker tasks to spawn.
-/// * `results_sender` - The channel to emit results (task and proof).
-/// * `prover_event_sender` - The channel to send prover events to the main thread.
-///
-/// # Returns
-/// A tuple containing:
-/// * A vector of `Sender<Task>` for each worker, allowing tasks to be sent to them.
-/// * A vector of `JoinHandle<()>` for each worker, allowing the main thread to await their completion.
-pub fn start_workers(
-    num_workers: usize,
-    results_sender: mpsc::Sender<(Task, Proof)>,
-    event_sender: mpsc::Sender<Event>,
-    shutdown: broadcast::Receiver<()>,
-    environment: Environment,
-    client_id: String,
-) -> (Vec<mpsc::Sender<Task>>, Vec<JoinHandle<()>>) {
-    let mut senders = Vec::with_capacity(num_workers);
-    let mut handles = Vec::with_capacity(num_workers);
-
-    for worker_id in 0..num_workers {
-        let (task_sender, mut task_receiver) = mpsc::channel::<Task>(8);
-        // Clone senders and receivers for each worker.
-        let prover_event_sender = event_sender.clone();
-        let results_sender = results_sender.clone();
-        let mut shutdown_rx = shutdown.resubscribe();
-        let client_id = client_id.clone();
-        let error_classifier = ErrorClassifier::new();
-        let handle = tokio::spawn(async move {
-            loop {
-                tokio::select! {
-                    _ = shutdown_rx.recv() => {
-                        let message = format!("Worker {} received shutdown signal", worker_id);
-                        let _ = prover_event_sender
-                            .send(Event::prover(worker_id, message, EventType::Shutdown))
-                            .await;
-                        break; // Exit the loop on shutdown signal
-                    }
-                    // Check if there are tasks to process
-                    Some(task) = task_receiver.recv() => {
-                        match authenticated_proving(&task, &environment, client_id.clone()).await {
-                            Ok(proof) => {
-                                let message = format!(
-                                    "Proof completed successfully (Prover {})",
-                                    worker_id
-                                );
-                                let _ = prover_event_sender
-                                    .send(Event::prover(worker_id, message, EventType::Success))
-                                    .await;
-                                let _ = results_sender.send((task, proof)).await;
-                            }
-                            Err(e) => {
-                                let log_level = error_classifier.classify_worker_error(&e);
-                                let message = format!("Error: {}", e);
-                                let event = Event::prover_with_level(worker_id, message, EventType::Error, log_level);
-                                if event.should_display() {
-                                    let _ = prover_event_sender.send(event).await;
-                                }
-
-                                // For analytics errors, continue processing but don't send result
-                                // For other errors, also don't send result (task failed)
-                            }
-                        }
-                    }
-                    else => break,
-                }
-            }
-        });
-
-        senders.push(task_sender);
-        handles.push(handle);
-    }
-
-    (senders, handles)
+    offline::start_anonymous_workers(num_workers, shutdown, environment, client_id).await
 }
 
 #[cfg(test)]
 mod tests {
     use crate::orchestrator::MockOrchestrator;
-    use crate::prover_runtime::{Event, MAX_COMPLETED_TASKS, fetch_prover_tasks};
+    use crate::prover_runtime::{Event, MAX_COMPLETED_TASKS, online::fetch_prover_tasks};
     use crate::task::Task;
     use crate::task_cache::TaskCache;
     use std::time::Duration;
