+++ clients/cli/src/workers/mod.rs
@@ -1,0 +1,2 @@
+pub mod offline;
+pub mod online;
